huggingface-llm-project/requirements.txt
huggingface-hub
huggingface-transformers
flask
transformers
torch
fastapi
uvicorn
python-dotenv